+++ webpack.config.js
@@ -33,7 +33,7 @@ module.exports = {
         use: ['to-string-loader', 'css-loader']
       },
       {
-        test: /\.(png|jpg|gif)$/i,
+        test: /\.(png|jpg|gif|mp3)$/i,
         use: [
           {
             loader: 'url-loader',